terminal-helper/requirements.txt
openai
tiktoken
duckduckgo-search
termcolor
python-dotenv
beautifulsoup4
lxml
colorama
